Beginning of our Report due this Wednesday
I have written a little bit that will contribute to my portion of the
report that is due this Wednesday.  It's just a little background on A*
in general, as I find that that should be useful.  Then I will discuss
what I intend to do to parallelize it (starting with parallelizing a
bidirectional search with A* - depending on what algorithms you guys are
doing, you can also speed things up by using a bidirectional search),
and maybe get to running some diagnostics, if I get that far.  I'm
pretty sure that with this Midterm Report it's ok if we don't have much
data yet on how things are going, just SOME code to show that we're on
the right track and have some semblance that we know what we're talking
about and have done some research.